moving get_host_system_profile_by_id to api.host
